Refactor: 홈 컴포넌트 index 사용으로 Import문 정리
