+++ app/page.tsx
@@ -1,8 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
-import Txt from "@/components/atoms/Text";
-import { Card, Olim, Profile, Savings } from "@/components/home";
-import Splash from "@/components/home/Splash";
+import { Txt } from "@/components/atoms";
+import { Card, Olim, Profile, Savings, Splash } from "@/components/home";
 import { getLetterCount } from "@/lib/actions/home-actions";
 import { requireAuth } from "@/utils/auth";
