+++ components/home/index.ts
@@ -1,4 +1,9 @@
+export { default as Letter } from "./hanaolim/Letter";
+export { default as Point } from "./hanaolim/Point";
+export { default as PointRuleTooltip } from "./hanaolim/PointRuleTooltip";
 export { default as Card } from "./Card";
 export { default as Olim } from "./Olim";
 export { default as Profile } from "./Profile";
+export { default as RocketSplash } from "./RocketSplash";
 export { default as Savings } from "./Savings";
+export { default as Splash } from "./Splash";
